Move mappers to mappers directory
