+++ src/mappers/aggregate-data-to-chart-data-mapper.js
